+++ docs/autoapi_templates/index.rst
@@ -72,6 +72,10 @@ All operators are in the following packages:
 
   airflow/gcp/sensors/index
 
+  airflow/providers/aws/operators/index
+
+  airflow/providers/aws/sensors/index
+
   airflow/providers/google/cloud/operators/index
 
   airflow/providers/google/marketing_platform/operators/index
@@ -99,6 +103,8 @@ All hooks are in the following packages:
   airflow/contrib/hooks/index
 
   airflow/gcp/hooks/index
+
+  airflow/providers/aws/hooks/index
 
   airflow/providers/google/cloud/hooks/index
