+++ docs/conf.py
@@ -210,6 +210,8 @@ exclude_patterns = [
     '_api/airflow/gcp/example_dags',
     '_api/airflow/gcp/utils',
     '_api/airflow/providers/index.rst',
+    '_api/airflow/providers/aws/index.rst',
+    '_api/airflow/providers/aws/example_dags',
     '_api/airflow/providers/google/index.rst',
     '_api/airflow/providers/google/cloud/index.rst',
     '_api/airflow/providers/google/cloud/example_dags',
